Adding support for float and adding data file

+++ HW1/src/CSVReader.py
@@ -33,7 +33,7 @@ class CSVReader(object):
 		self.data_types = []
 		for i in range(len(self.columns)):
 			if self.columns[i][0] == '$' or self.columns[i][0] == '<' or self.columns[i][0] == '>':
-				self.data_types.append(int)
+				self.data_types.append(float)
 			else:
 				self.data_types.append(str)
